% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_info_doc.R
\docType{data}
\name{dataset_info}
\alias{dataset_info}
\title{Dataset information}
\format{
An object of class \code{list} of length 17.
}
\usage{
dataset_info = list(
 
     #file string parts names
     temp ="tas", 
     prec ="pr", 
     lwnet ="lwnet", 
     rsds ="rsds", 
     soil ="soil", 
     wet ="wet", 
     co2 = "CO2",
     
     #Shared variables
     wyears = seq(1951,2010,5), 
     cells = 67420, 
     
     #readLPJ specific variables more information on 
     \code{\link{readLPJ}}
     
     file_name = "pft_harvest.pft.bin",
     crop ="mgrass",
     water = "rainfed",
     syear = 1951,
     years = 149,
     bands = 32,
     soilcells = FALSE, #True (59199 cells), False (67420 cells)
     monthly = FALSE
)
}
\arguments{
\item{temp}{String that identify the temperature file}

\item{prec}{String that identify the temperature file}

\item{lwnet}{String that identify the temperature file}

\item{rsds}{String that identify the temperature file}

\item{soil}{String that identify the temperature file}

\item{wet}{String that identify the temperature file}

\item{co2}{String that identify the temperature file}

\item{wyears}{vector of years that will be extracted from}

\item{cells}{number os cells used in the simulation}

\item{file_name}{output name of the file outcome of the Lpjml runs}

\item{crop}{crop name}

\item{water}{information about the water avaialability, either \code{irrigation} or \code{rainfed}}

\item{syear}{LPJml simulations starting year}

\item{years}{total number of simulated years (usually syear-2100)}

\item{bands}{The number of bands (e.g. crops, months). Will be 12 if monthly==TRUE}

\item{soilcells}{Bool. If 67420 cells are present in the file, should only the 59199 for MAgPIE be returned?}

\item{monthly}{\code{true} for monthly data and \code{false} for other intervals}
}
\description{
This list called \code{dataset_info} aims to be an oobject-like variable that stores general information used to read and process binary 
inputs and outputs from LPJmL. For consistency, it is advisible that this list be created only once and loaded in all functions that need
need it as an input (\code{\link{lpjml_inputs}}, \code{\link{lpjml_outputs}}, \code{\link{df_trainig}})
}
\keyword{datasets}
