% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\docType{data}
\name{grid}
\alias{grid}
\title{grid file}
\format{
A data frame with columns:
\describe{
 \item{lon}{longitude}
 \item{lat}{latitute}
}
}
\source{
lpjml 3.0 grid file information
}
\usage{
grid
}
\description{
grid file
}
\examples{
\dontrun{
 grid
}
}
\keyword{datasets}
