% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjml_inputs2.R
\name{lpjml_inputs2}
\alias{lpjml_inputs2}
\title{LPjmL input preparation}
\usage{
lpjml_inputs2(folder, plotting = T, dataset_info)
}
\arguments{
\item{folder}{Folder path for the input files}

\item{plotting}{\code{True} or \code{False} determing if plots from the data extracted will be plotted for checking.}

\item{dataset_info}{object list containing information about the files that will be used that are going to be extracted.}
}
\description{
Function `lpjml_output` read binary with the output of LPJmL runs and processes them into `dataframes`.
}
\details{
This function must work with a folder organiyation of the output files of LPJmL
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
