% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjm_output.R
\name{lpjml_output}
\alias{lpjml_output}
\title{LPjmL output preparation}
\usage{
lpjml_output(folder, datasets_info)
}
\arguments{
\item{datasets_info}{object list containing information about the files that will be used that are going to be extracted.
for more information plese check \link{dataset_info}.}
}
\description{
Function `lpjml_output` read binary with the output of LPJmL runs and processes them into `dataframes`.
}
\details{
This function must work with a folder organiyation of the output files of LPJmL
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
