% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjm_output2.R
\name{lpjml_output2}
\alias{lpjml_output2}
\title{LPjmL output preparation}
\usage{
lpjml_output2(folder, dataset_info, rel)
}
\arguments{
\item{folder}{Folder path for the input files
for more information plese check @Run \code{vignette("lpjmule", package = "lpjmule")} .}

\item{dataset_info}{object list containing information about the files that will be used that are going to be extracted.}
}
\description{
Function `lpjml_output` read binary with the output of LPJmL runs and processes them into `dataframes`.
}
\details{
This function must work with a folder organiyation of the output files of LPJmL
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
