% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjmule-package.R
\docType{package}
\name{lpjmule-package}
\alias{lpjmule}
\alias{lpjmule-package}
\title{lpjmule: This package reads LPJmL inputs and outputs to create an emulator of of LPJmL that can be emebeded in MAgPIE}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\details{
This oackage gatter functions to read and process lpjml input and output files into machine learning models built with keras for the purpose of 
creating an emulator of LPJmL that can be used in MAgPIE.
}
\keyword{internal}
