% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{magclassdata}
\alias{magclassdata}
\title{magclassdata system files}
\format{
An object of class \code{list} of length 6.
}
\source{
magclass package
}
\usage{
magclassdata
}
\description{
magclassdata system files
}
\examples{
\dontrun{
 magClassdata
}
}
\keyword{datasets}
