% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_warper.R
\name{training_warper}
\alias{training_warper}
\title{Machine learning training dataset}
\usage{
training_warper(folder, input_levels = c(seq(0, 2, 0.2), 2.5), dataset_info)
}
\arguments{
\item{folder}{Folder path for the input and output files}

\item{input_levels}{The breakdown of the LSU leves in the data ser as a vector.}

\item{dataset_info}{Object list containing information about the files that will be used that are going to be extracted.}
}
\description{
Function `training_warper` calls functions `lpjml_output` and `Lpjml_input` and `df_training` as appropriate to create and save the training dataset.
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
