% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_training.R
\name{df_training}
\alias{df_training}
\title{Machine learning training dataset}
\usage{
df_training(
  lpjml_inputs,
  lpjml_output,
  input_levels = c(seq(0, 2, 0.2), 2.5),
  dataset_info
)
}
\arguments{
\item{lpjml_inputs}{input from function `Lpjml_input`}

\item{lpjml_output}{output from function `lpjml_output`}

\item{input_levels}{The breakdown of the LSU leves in the data ser as a vector.}

\item{dataset_info}{object list containing information about the files that will be used that are going to be extracted.}
}
\description{
Function `df.training` combine together the output of the `lpjml_output` and `Lpjml_input` functions into a `features:label` format.
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
