% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_info_doc.R
\docType{data}
\name{grid}
\alias{grid}
\title{Applicants.}
\format{
A data frame with 3 variables: \code{year}, \code{sex} and
  \code{grid}
}
\usage{
grid
}
\description{
Information about lpjml grid used for checking the if the input data was processed correctly
}
\keyword{datasets}
