% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpjm_output.R
\name{lpjml_output}
\alias{lpjml_output}
\title{LPjmL output preparation}
\usage{
lpjml_output(
  folder,
  wyears = seq(1951, 2010, 10),
  cells = 67420,
  file_name = "pft_harvest.pft.bin",
  crop = "mgrass",
  water = "rainfed"
)
}
\arguments{
\item{folder}{Folder path for the input files}

\item{wyears}{Sequence of years to extracted}

\item{cells}{Number of grid cells (67420)}

\item{file_name}{Name of the binary files containing LPJmL results}

\item{crop}{Name of the crop to be extracted}

\item{water}{Source of water, eiter `rainfed` or `irrigated`. Currently only `rainfed` is available}
}
\description{
Function `lpjml_output` read binary with the output of LPJmL runs and processes them into `dataframes`.
}
\details{
This function must work with a folder organiyation of the output files of LPJmL
}
\author{
Marcos Alves \email{mppalves@gmail.com}
}
